Testing APIs for historical weather data retrieval
Multiple APIs were tested and finally the open-meteo API was settled upon. An example retrieving data for 17 days has been added for testing and values in there will be dynamically changed to  provide input to the model.
Dependencies:
svelte-jsoneditor was added for easier viewing of the JSON data the API returns.
open-meteo was added to retrieve data by using the Open-Meteo API.